==> 5_Merging_Summary_Tables.ipynb <==
# Pandas is a package containing additional functions to use data frames in Python
import pandas as pd

# These two lines allow the notebook to access the Google Drive.
from google.colab import drive
drive.mount('/content/drive', force_remount=True)

# This is the path to the project folder within the Google Drive.
file_path = "/content/drive/My Drive/"

---
## Notebook 5

# Merging Summary Tables

This notebook contains the steps to merge the outputs of notebook 2 - Data Cleaning Automated - and notebook 4 - Georeferencing Automated - and a manually generated basic information table into a combined summary table.

data_cleaning = pd.read_csv(file_path + "final_summary_data_cleaning.tsv", sep="\t")
georeferencing = pd.read_csv(file_path + "final_summary_georeferencing.tsv", sep="\t")
wiki = pd.read_csv(file_path + "basic_info_wikipedia.tsv", sep="\t")

combined_final_summary_table = data_cleaning.merge(georeferencing)
combined_final_summary_table

final_summary_all = combined_final_summary_table.merge(wiki)
final_summary_all

saving final_summary_all to google drive:

final_summary_all.to_csv(file_path + "/final_summary_all_tables/final_summary_all.tsv", sep="\t", index=None)

final_summary_all[final_summary_all["family_wiki"] != test["family"]]

filt = final_summary_all[final_summary_all['nrecords_final'] < 50]
filt = filt[filt['nrecords_unfiltered'] > 50]
filt

filt.to_csv(file_path + "final_summary_all_tables/filtered_final_summary_all.tsv", sep="\t", index=None)
